# backend/api/build/resources/main/openapi/dopashift-api-v1.yml
openapi: 3.0.3
info:
  title: DopaShift API
  description: |
    REST API for the DopaShift multi-goal life-tracking platform.
    
    ## Versioning Policy
    - All endpoints are prefixed with `/v1/`
    - Previous API versions are supported for 6 months after a new version is released.
    - Breaking changes will only be introduced in new major versions.
    
    ## Authentication
    All endpoints (except /v1/auth/*) require a valid Bearer token in the Authorization header.
    Tokens are obtained via the Keycloak-proxied authentication endpoints.
  version: 1.0.0
  contact:
    name: DopaShift Team
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api.dopashift.app
    description: Production
security:
  - bearerAuth: []
...
paths:
...
  /v1/goals:
    post:
      tags: [Goals]
      operationId: createGoal
      summary: Create a new goal profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGoalRequest'
      responses:
        '201':
          description: Goal created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GoalResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Goal name already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      tags: [Goals]
      operationId: listGoals
      summary: List all goals for the authenticated user
      responses:
        '200':
          description: List of goals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GoalResponse'
...
  /v1/goals/{id}/checklist:
    post:
      tags: [Checklist]
      operationId: createChecklistItem
      summary: Add a checklist item to a goal
      parameters:
        - $ref: '#/components/parameters/GoalId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateChecklistItemRequest'
      responses:
        '201':
          description: Checklist item created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChecklistItemResponse'
    get:
      tags: [Checklist]
      operationId: listChecklistItems
      summary: List checklist items for a goal
      parameters:
        - $ref: '#/components/parameters/GoalId'
      responses:
        '200':
          description: Checklist items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChecklistItemResponse'
  /v1/goals/{goalId}/checklist/{itemId}:
    put:
      tags: [Checklist]
      operationId: updateChecklistItem
      parameters:
        - name: goalId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: itemId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateChecklistItemRequest'
      responses:
        '200':
          description: Item updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChecklistItemResponse'
    delete:
      tags: [Checklist]
      operationId: deleteChecklistItem
      parameters:
        - name: goalId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: itemId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Item deleted

  # === Todos ===
  /v1/todos:
    post:
      tags: [Todos]
      operationId: createTodo
      summary: Create a daily todo item (max 100/day)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTodoRequest'
      responses:
        '201':
          description: Todo created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          description: Daily limit reached (100 items)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      tags: [Todos]
      operationId: listTodos
      summary: List todos for a specific date
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
          example: "2024-06-15"
      responses:
        '200':
          description: Todos for the date
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TodoResponse'
  /v1/todos/{id}:
    put:
      tags: [Todos]
      operationId: updateTodo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTodoRequest'
      responses:
        '200':
          description: Todo updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoResponse'
    delete:
      tags: [Todos]
      operationId: deleteTodo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Todo deleted
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    GoalId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [code, message, correlationId, timestamp]
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: "Goal name must be 1-100 characters"
        details:
          type: object
          additionalProperties: true
        correlationId:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time

    # === Auth ===
...
    CreateGoalRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        keywords:
          type: array
          items:
            type: string
            minLength: 1
            maxLength: 50
          maxItems: 20
...
    GoalResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        keywords:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    CreateChecklistItemRequest:
      type: object
      required: [text]
      properties:
        text:
          type: string
          maxLength: 500
    UpdateChecklistItemRequest:
      type: object
      properties:
        text:
          type: string
          maxLength: 500
        isCompleted:
          type: boolean
    ChecklistItemResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        goalId:
          type: string
          format: uuid
        text:
          type: string
        isCompleted:
          type: boolean
        createdAt:
          type: string
          format: date-time

    # === Todos ===
    CreateTodoRequest:
      type: object
      required: [text]
      properties:
        text:
          type: string
          minLength: 1
          maxLength: 500
        dayDate:
          type: string
          format: date
        dueDateTime:
          type: string
          format: date-time
    UpdateTodoRequest:
      type: object
      properties:
        text:
          type: string
          minLength: 1
          maxLength: 500
        isCompleted:
          type: boolean
    TodoResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        text:
          type: string
        isCompleted:
          type: boolean
        dayDate:
          type: string
          format: date
        dueDateTime:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # === Habits ===
...